requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
arabic-reshaper==3.0.0
arrow==1.3.0
attrs==23.2.0
binaryornot==0.4.4
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cookiecutter==2.6.0
cryptography==46.0.3
dataclasses-json==0.6.6
defusedxml==0.7.1
distro==1.9.0
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.110.0
firebase-admin==6.5.0
flet==0.24.1
flet-core==0.24.1
flet-runtime==0.24.1
fonttools==4.53.1
fpdf2==2.7.9
frozenlist==1.4.1
gcloud==0.18.3
google-api-core==2.17.1
google-api-python-client==2.122.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.15.0
google-cloud-storage==2.15.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio>=1.62.1
grpcio-status>=1.62.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.1
httpx<0.28
idna==3.6
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jws==0.1.3
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.3
langchain-openai==0.1.8
langchain-text-splitters==0.2.0
langsmith==0.1.65
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
mdurl==0.1.2
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
openai==1.30.5
orjson==3.10.3
packaging==23.2
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
proto-plus==1.23.0
protobuf==4.25.8
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.6.3
pydantic_core==2.16.3
Pygments==2.17.2
PyJWT==2.8.0
PyMuPDF>=1.24.0
pyparsing==3.1.2
pypdfium2==4.30.0
pypng==0.20220715.0
Pyrebase4==4.7.1
python-bidi==0.6.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-http-client==3.3.7
python-jwt==2.0.1
python-slugify==8.0.4
PyYAML==6.0.1
qrcode==7.4.2
regex==2024.5.15
repath==0.9.0
requests==2.29.0
requests-toolbelt==0.10.0
rich==13.7.1
rsa==4.9
six==1.16
sniffio==1.3.1
SQLAlchemy==2.0.30
starkbank-ecdsa==2.2.0
starlette==0.36.3
tenacity==8.3.0
text-unidecode==1.3
tiktoken==0.7.0
tqdm==4.66.4
twilio==9.0.0
types-python-dateutil==2.8.19.20240311
typing_extensions==4.10.0
typing-inspect==0.9.0
uritemplate==4.1.1
urllib3==1.26.18
uvicorn==0.28.0
watchdog>=2.1.0
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4
setuptools<82
